+++ tests/unit/cli/output/helpers/ProgressBarTest.php
@@ -183,4 +183,56 @@ class ProgressBarTest extends TestCase
 			$progressBar->advance();
 		}
 	}
+
+	/**
+	 *
+	 */
+	public function testRemoveIncomplete()
+	{
+		$output = Mockery::mock('mako\cli\output\Output');
+
+		$output->shouldReceive('write')->once()->with("\r00/10 --------------------   0% ");
+
+		$output->shouldReceive('clearLines')->once()->with(1);
+
+		$progressBar = new ProgressBar($output, 10);
+
+		$progressBar->draw();
+
+		$progressBar->remove();
+	}
+
+	/**
+	 *
+	 */
+	public function testRemoveComplete()
+	{
+		$output = Mockery::mock('mako\cli\output\Output');
+
+		$output->shouldReceive('write')->once()->with("\r00/10 --------------------   0% ");
+		$output->shouldReceive('write')->once()->with("\r01/10 ==------------------  10% ");
+		$output->shouldReceive('write')->once()->with("\r02/10 ====----------------  20% ");
+		$output->shouldReceive('write')->once()->with("\r03/10 ======--------------  30% ");
+		$output->shouldReceive('write')->once()->with("\r04/10 ========------------  40% ");
+		$output->shouldReceive('write')->once()->with("\r05/10 ==========----------  50% ");
+		$output->shouldReceive('write')->once()->with("\r06/10 ============--------  60% ");
+		$output->shouldReceive('write')->once()->with("\r07/10 ==============------  70% ");
+		$output->shouldReceive('write')->once()->with("\r08/10 ================----  80% ");
+		$output->shouldReceive('write')->once()->with("\r09/10 ==================--  90% ");
+		$output->shouldReceive('write')->once()->with("\r10/10 ==================== 100% ");
+		$output->shouldReceive('write')->once()->with(PHP_EOL);
+
+		$output->shouldReceive('clearLines')->once()->with(2);
+
+		$progressBar = new ProgressBar($output, 10);
+
+		$progressBar->draw();
+
+		for($i = 0; $i < 10; $i++)
+		{
+			$progressBar->advance();
+		}
+
+		$progressBar->remove();
+	}
 }